+++ closet-infra/src/main/java/project/closet/api/kakaoapi/KakaoAddressClient.java
@@ -1,4 +1,4 @@
-package project.closet.api;
+package project.closet.api.kakaoapi;
 
 import io.github.bucket4j.Bucket;
 import java.util.List;
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.client.RestTemplate;
+import project.closet.api.AddressClient;
 import project.closet.api.response.KakaoAddressResponse;
 import project.closet.api.response.KakaoAddressResponse.Document;
 
@@ -30,7 +31,8 @@ public class KakaoAddressClient implements AddressClient {
     public KakaoAddressResponse requestAddressFromKakao(Double longitude, Double latitude) {
 
         if (!kakaoApiBucket.tryConsume(1)) {
-            throw new RuntimeException();
+            log.warn("[RateLimit] Kakao API 일일 또는 월간 호출 한도 초과");
+            throw new RateLimitExceededException("Kakao API 호출 한도를 초과했습니다.");
         }
 
         String url =
