+++ closet-infra/src/main/java/project/closet/api/kakaoapi/RateLimitExceededException.java
@@ -1,0 +1,12 @@
+package project.closet.api.kakaoapi;
+
+public class RateLimitExceededException extends RuntimeException {
+
+    public RateLimitExceededException(String message) {
+        super(message);
+    }
+
+    public RateLimitExceededException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
